% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{clean_activity_df}
\alias{clean_activity_df}
\title{Get Clean Activity Data}
\usage{
clean_activity_df(
  strava_df,
  peloton_df = NULL,
  other = c("Stretching", "Cardio", "Meditation")
)
}
\arguments{
\item{strava_df}{df output of clean_strava_file. required.}

\item{peloton_df}{df output of clean_peloton_file. optional}

\item{other}{vector of strings that are other workouts}
}
\value{
data frame with single entry for each activity and a group
}
\description{
Combines Strava and Peloton data into a single data set that avoids
duplicating workouts shared from Peloton to Strava. Will work if no peloton
file is had, creating grouped data.
}
