% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{clean_monthly_type_summary}
\alias{clean_monthly_type_summary}
\title{Make Daily Activities (with distance) by Month Data}
\usage{
clean_monthly_type_summary(df, typecolors = NULL)
}
\arguments{
\item{df}{data.frame out of clean_activity_df. could be created from
another source. Is expected to have columns: c('date','month',
'type','duration','pace','distance','elev')}

\item{typecolors}{data.frame of colors by type.
defaults to activity_type_colors()}
}
\value{
data.frame summarized
}
\description{
Used in plots
}
\examples{
\dontrun{
# use the example data
clean_monthly_type_summary(df = ex_actv_df)
}
}
