% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton.R
\name{clean_peloton_file}
\alias{clean_peloton_file}
\title{Clean Peloton Data from File}
\usage{
clean_peloton_file(fpath, this_year = NULL, prefer_utc = FALSE, tzs = NULL)
}
\arguments{
\item{fpath}{path to the csv}

\item{this_year}{numeric year defaults to NULL
if NULL there is no filter}

\item{prefer_utc}{boolean for if the timestamp should be
converted to UTC. default is FALSE (local at time of activity)
conversion to UTC has challenges, namely that three digit
timezones are not unique (CST used in US, China, and sometimes Oz)}

\item{tzs}{list of three character timezones
that you should definitely pass if you know what your data
have in them}
}
\value{
dataframe
}
\description{
Clean Peloton Data from File
}
\examples{
\dontrun{
peloton_df <- clean_peloton_file(fpath = file.path("dir", "peloton.csv"),
                                 this_year = year(Sys.Date())-1)
}
}
