% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava.R
\name{clean_strava_file}
\alias{clean_strava_file}
\title{Clean Strava File}
\usage{
clean_strava_file(fpath, this_year = NULL, this_tz = NULL, has_peloton = FALSE)
}
\arguments{
\item{fpath}{path to the activities.csv from your strava download. Strava
gives you an entire folder of data when you download your data. The path
just to the activities.csv is desired here.}

\item{this_year}{numeric. like 2025.}

\item{this_tz}{string. like UTC. Only use this if all of your activities are
from the same time zone that you know.}

\item{has_peloton}{boolean. if you want to label rides with zero elevation
as peloton ride.}
}
\value{
data frame of strava activites
}
\description{
This is an opinionated function using Imperial units, returning temperature
in F, distance in miles, and duration in minutes.
}
\examples{
\dontrun{
strava_df <- clean_strava_file(fpath = file.path("dir", "activities.csv"),
                               this_year = year(Sys.Date())-1)
}
}
