% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton.R
\name{decode_timezones}
\alias{decode_timezones}
\title{Decode timezones}
\usage{
decode_timezones(x, tzs = NULL)
}
\arguments{
\item{x}{vector of three character timezones}

\item{known_timezones}{list of three character timezones
that you should definitely pass if you know what your data
have in them}
}
\description{
Decode timezones
}
