% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton.R
\docType{data}
\name{known_timezones}
\alias{known_timezones}
\title{Known timezones}
\format{
An object of class \code{list} of length 17.
}
\usage{
known_timezones
}
\description{
Known timezones
}
\keyword{datasets}
