% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_daily_activities_by_month_plot}
\alias{make_daily_activities_by_month_plot}
\title{Make Daily Activities by Month Plot}
\usage{
make_daily_activities_by_month_plot(actv_df, this_year, groupcolors = NULL)
}
\arguments{
\item{actv_df}{data.frame out of clean_activity_df. could be created from
another source. Is expected to have columns: c('date','month','name',
'type','duration','pace','distance','elev','condition','temp','group')}

\item{this_year}{numeric. like 2025}

\item{groupcolors}{data.frame of colors by group}
}
\value{
plot
}
\description{
Creates a horizontal plot with a breakdown for each month with number of
activities and summarizes by duration
}
\examples{
\dontrun{
# use the example data
make_daily_activities_by_month_plot <- function(ex_actv_df, 2025)
}
}
