% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_mo_elev_type_plot}
\alias{make_mo_elev_type_plot}
\title{Make Monthly Elevation by Type Plot}
\usage{
make_mo_elev_type_plot(df, typecolors = NULL)
}
\arguments{
\item{df}{data.frame expected to have columns: c('month', 'type'
'elev'). output of clean_monthly_type_summary() is best.}

\item{typecolors}{data.frame of colors by type}
}
\value{
plot
}
\description{
Creates a bar chart of vert by type
}
\examples{
\dontrun{
# use the example data
mo_df <- clean_monthly_type_summary(df = ex_actv_df)
make_mo_elev_type_plot(df = mo_df)
}
}
