% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton.R
\name{replace_timezones_in_timestamps}
\alias{replace_timezones_in_timestamps}
\title{Replace timezones in timestamps}
\usage{
replace_timezones_in_timestamps(x, tzs = NULL)
}
\arguments{
\item{x}{vector of timestamps}

\item{tzs}{list of three character timezones
that you should definitely pass if you know what your data
have in them}
}
\description{
Replace timezones in timestamps
}
